debug on output

+++ SnifferCxx/model/Coordinate.h
@@ -16,7 +16,7 @@
 namespace Model {
 	class WindVector;
 
-	typedef int coord_item_t;
+	typedef long coord_item_t;
 	typedef std::array<coord_item_t, 3> coord_t;
 
 	class Coordinate : public coord_t {
